refactors playground to features
