+++ web/src/features/playground/client/components/PromptVariableComponent.tsx
